Consolidate neo4j query code in a single file for now, as there was no reason to keep it in more than one.
