+++ ontology/neo4j/concept.go
